Create stripped down db creation.yag
This command is the same as `thread db mgmt - stripped down version.yag` EXCEPT the code is much easier to read and the database entry structure is different.
This is not compatible with anything in the parent folder. To use things like the emoji addon, you'd have to update the db values it checks for.